move ValidatorContent to Generator
